test(parser/hwp3): 라틴 확장 디코딩 계약을 tests/cases 통합 테스트로 이관 (#5555)
src/ 안 단위 테스트 증가는 테스트 매니페스트 래칫(#4976)이 금지한다.
johab.rs::tests 의 신규 테스트를 tests/cases 로 옮기고, 실측 8문자 항등과
구간 밖 사적 따옴표 유지를 두 계약으로 나눠 검사한다.
